consolidate DataViews subcomponent export to reduce changes
