+++ packages/dataviews/src/components/dataviews/stories/index.story.tsx
@@ -16,7 +16,7 @@ import { __, _n } from '@wordpress/i18n';
 /**
  * Internal dependencies
  */
-import DataViews from '../../';
+import DataViews from '../index';
 import {
 	DEFAULT_VIEW,
 	actions,
