+++ packages/dataviews/src/index.ts
@@ -1,4 +1,4 @@
-export { default as DataViews } from './components';
+export { default as DataViews } from './components/dataviews';
 export { default as DataForm } from './components/dataform';
 export { VIEW_LAYOUTS } from './dataviews-layouts';
 export { filterSortAndPaginate } from './filter-and-sort-data-view';
